Adding very very basic tests
