+++ tsconfig.json
@@ -27,7 +27,6 @@
     "scripts",
     "acceptance-tests",
     "webpack",
-    "jest",
     "src/setupTests.ts",
     "pipeline-scripts"
   ]
